fix: move auth folder into the correct one following the correct architecture of the project
