+++ src/modules/users/users.controller.ts
@@ -9,7 +9,7 @@ import {
   BadRequestException,
 } from '@nestjs/common';
 import { UsersService } from './services/users.service';
-import { Public } from '../../auth/decorators/public.decorator';
+import { Public } from '../../modules/auth/decorators/public.decorator';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { DriveService } from '../drive/drive.service';
